<3 low level memory managment

+++ include/Containers/BitSet.cuh
@@ -75,7 +75,7 @@ bool BitSet::contains(u32 value) const
 __host__ __device__
 void BitSet::clear()
 {
-    thrust::fill(thrust::seq, begin(), end(), 0);
+    thrust::fill(thrust::seq, begin(), end(), static_cast<u32>(0));
 }
 
 __host__ __device__
